kpa_assignment/requirements.txt
Django==5.2.4
djangorestframework
mysqlclient
django-cors-headers
gunicorn